requirements.txt
colorlog
httpx[http2]
oauthlib
requests_oauthlib
attrs